added the initial feature of notes

+++ cities/src/main/java/com/youssef/library/cities/DTOs/Review/ReviewInsertion/ReviewInsertionDTO.java
@@ -9,5 +9,5 @@ import lombok.NoArgsConstructor;
 @NoArgsConstructor
 public class ReviewInsertionDTO {
     private String review;
-    private double stars;
+    private int stars;
 }
